EES-5135 Move `DataSetVersion` view models into separate files
